requirements.txt
selenium-stealth
pyyaml
requests
phonenumbers